fix: schema
